pyproject.toml
[project]
name = "schulpipeline"
version = "0.1.0"
description = "Spec-driven pipeline for automating school assignments"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "python-pptx>=1.0",
    "python-docx>=1.1",
    "beautifulsoup4>=4.12",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]
ollama = []  # no extra deps, just a config flag

[project.scripts]
schulpipeline = "schulpipeline.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
